make check_panic and snapshot_diagnostics standalone functions
